% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ci.R
\name{add_ci}
\alias{add_ci}
\title{Add confidence interval for the indirect effect to a fitted mediation model}
\usage{
add_ci(mediation_model, ...)
}
\arguments{
\item{mediation_model}{A mediation model fitted with \code{mediation_model}
method.}

\item{...}{Further argument to pass to the different methods.}

\item{iter}{Number of iteration for the Monte Carlo CI method.}

\item{alpha}{Alpha threshold to use when computing Monte Carlo CI.}
}
\value{
An object of the same class as \code{mediation_model}, but with
confidence interval for the indirect effect estimation added for later
use.
}
\description{
Add confidence interval for the indirect effect to a fitted mediation model
}
\examples{
\dontrun{
mediation_model <- add_ci(mediation_model)
}
}
