% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ci.R
\name{add_index}
\alias{add_index}
\title{Add index for the indirect effect to a fitted mediation model}
\usage{
add_index(mediation_model, ...)
}
\arguments{
\item{mediation_model}{A mediation model fitted with \code{mediation_model}
method.}

\item{CI}{a boolean indecating if CI should be computed. Defaults to
\code{FALSE}.}

\item{iter}{number of iteration for the Monte Carlo CI method.}

\item{alpha}{Alpha threshold to use when computing Monte Carlo CI.}

\item{stage}{Stage on which you want to compute conditional index for
moderated mediation models.}
}
\value{
An object of the same class as \code{mediation_model}, dbut with
confidence interval for the indirect effect estimation added for later
use.
}
\description{
Add index for the indirect effect to a fitted mediation model
}
\examples{
\dontrun{
mediation_model <- add_ci(mediation_model)
}
}
