% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_simple_index.R
\name{add_index.simple_mediation}
\alias{add_index.simple_mediation}
\title{add_index method for simple mediation}
\usage{
\method{add_index}{simple_mediation}(mediation_model, iter = 5000,
  alpha = 0.05, ...)
}
\arguments{
\item{mediation_model}{A mediation model of class
\code{"simple_mediation"}.}

\item{iter}{Number of simulation to use to compute Monte Carlo indirect
effect confidence interval.}

\item{alpha}{Alpha threshold to use with the confidence interval.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Adds confidence interval for the inderct effect to a mediation
model fitted with \code{\link{mdt_simple}}.
}
\examples{
## getting an indirect effect index
ho_et_al$condition_c <- build_contrast(ho_et_al$condition,
                                       "Low discrimination",
                                       "High discrimination")
simple_model <- mdt_simple(data = ho_et_al,
                           IV = condition_c,
                           DV = hypodescent,
                           M = linkedfate) 
add_index(simple_model) 
                                 
}
