% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_within_index.R
\name{add_index.within_participant_mediation}
\alias{add_index.within_participant_mediation}
\title{add_index method for within-participant mediation}
\usage{
\method{add_index}{within_participant_mediation}(mediation_model, iter = 5000,
  alpha = 0.05, ...)
}
\arguments{
\item{mediation_model}{A mediation model of class
\code{"within_participant_mediation"}.}

\item{iter}{Number of simulation to use to compute Monte Carlo indirect
effect confidence interval.}

\item{alpha}{Alpha threshold to use with the confidence interval.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Adds confidence interval for the inderct effect to a
within-participant mediation model fitted with \code{\link{mdt_within}} or
\code{\link{mdt_within_w}}.
}
\examples{
## getting an indirect effect index
within_model <- mdt_within(data = dohle_siegrist,
                           IV = name, 
                           DV = willingness, 
                           M = hazardousness, 
                           grouping = participant) 
add_index(within_model)

}
