% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_contrast.R
\name{build_contrast}
\alias{build_contrast}
\title{build a contrast code from character vector}
\usage{
build_contrast(vector, cond_a, cond_b)
}
\arguments{
\item{vector}{A character vector.}

\item{cond_a}{A character string to be coded -0.5}

\item{cond_b}{A character string to be coded  0.5}
}
\value{
A numeric vector.
}
\description{
This helper function helps you build a contrast code from a
character variable.
}
\examples{
  data(linkedfate)

  linkedfate$condition_contrast <- build_contrast(linkedfate$condition,
                                                  "Low discrimination",
                                                  "High discrimination")

 head(linkedfate)
}
