% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_assumptions.R
\name{check_assumptions}
\alias{check_assumptions}
\title{Test assumptions for models underlying the mediation}
\usage{
check_assumptions(mediation_model)
}
\arguments{
\item{mediation_model}{An object of class \code{mediation_model}.}
}
\value{
An object of class \code{mediation_model}.
}
\description{
When conducting a joint-significant test, different models are
fitted to the data. This function tests the normality and
heteroscedasticity assumptions for these models using \code{performance}.
\code{check_assumptions} uses \code{\link[performance]{check_normality}}
and \code{\link[performance]{check_heteroscedasticity}} to test the
assumptions.

Note that \code{check_assumptions} returns a
\code{mediation_model} object.
}
\examples{
data(ho_et_al)
ho_et_al$condition_c <- build_contrast(ho_et_al$condition,
                                       "Low discrimination",
                                       "High discrimination")
my_model <-
  mdt_simple(data = ho_et_al,
             IV = condition_c,
             DV = hypodescent,
             M = linkedfate)

check_assumptions(my_model)

}
