% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_indirect_effect_for.R
\name{compute_indirect_effect_for}
\alias{compute_indirect_effect_for}
\title{Compute the indirect effect index for a specific value of the moderator}
\usage{
compute_indirect_effect_for(
  mediation_model,
  Mod = 0,
  times = 5000,
  level = 0.05
)
}
\description{
@param Mod The moderator value for which to compute the indirect effect.
}
\details{
Note that the function relies on a model specification as described in Muller
et al. (2005).
}
