% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Example dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1312 rows and 3 columns.}
\usage{
dataset
}
\description{
Example dataset
}
\keyword{datasets}
