% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dohle_siegrist.R
\docType{data}
\name{dohle_siegrist}
\alias{dohle_siegrist}
\title{Data set showing within-subject analysis}
\format{A data frame with 44 rows and 4 variables:
\describe{
\item{subject}{Subject number.}
\item{name}{Name of the drugs ("simple" vs. "complex").}
\item{hazardousness}{Mean estimated hazardousness.}
\item{willingness}{Mean willigness to buy.}
}}
\usage{
data("dohle_siegrist")
}
\description{
A data set containing data from Experiment 1 from Dohle and
Siegrist (2014).
}
\references{
Dohle, S., & Siegrist, M. (2014). Fluency of pharmaceutical drug
names predicts perceived hazardousness, assumed side effects and willingness
to buy. \emph{Journal of Health Psychology}, \emph{19}(10), 1241‑1249.
doi: 10.1177/1359105313488974
}
\keyword{datasets}
