% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSmediation.R
\docType{package}
\name{JSmediation-package}
\alias{JSmediation}
\alias{JSmediation-package}
\title{JSmediation}
\description{
A set of helper functions to conduct mediation alaysis using
the joint significance approach and tidy eval.
}
\author{
\strong{Maintainer}: Cedric Batailler \email{cedric.batailler@univ-grenoble-alpes.fr}

Authors:
\itemize{
  \item Dominique Muller \email{dominique.muller@univ-grenoble-alpes.fr}
  \item Vincent Yzerbyt \email{vincent.yzerbyt@uclouvain.be}
  \item Charles Judd \email{charles.judd@colorado.edu}
}

}
