% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_moderated.R
\name{mdt_moderated}
\alias{mdt_moderated}
\title{Fit a moderated mediation model}
\usage{
mdt_moderated(data, IV, DV, M, Mod)
}
\arguments{
\item{data}{A data frame containing the variables in the model.}

\item{IV}{An unquoted variable in the data frame which will be used
as independant variable.}

\item{DV}{An unquoted variable in the data frame which will be used
as dependant variable.}

\item{M}{An unquoted variable in the data frame which will be used
as mediator.}

\item{Mod}{An unquoted variable in the data frame which will be used
as moderator.}
}
\value{
Returns an object of class "\code{mediation_model}" object.

  An object of class "\code{mediation_model}" is a least containing at list
  the components:

  \item{type}{A character string containing the type of model that has been
    conducted (e.g., \code{"simple mediation"}).}
  \item{method}{A character string containing the approach that has been
    used to conduct the mediation analysis (usually
    \code{"joint significance"}).}
  \item{params}{A named list of character strings describing variables used 
    in the model.}
  \item{paths}{A named list containing information on each relevent path of 
    the mediation model.}
  \item{indirect_index}{A boolean indicating wether indirect effect index has
    been computed. Defaults to \code{FALSE}. See \code{\link{add_index}} to
    compute mediation index.}
  \item{indirect_index_infos}{(Optional) An object of class
    \code{"indirect_index"}. Appears when one apply \code{\link{add_index}}
    to an object of class \code{"mediation_model"}.}
  \item{js_models}{A list of objects of class \code{"lm"}. Contains every
    model used in joint-significance test.}
  \item{data}{The original data frame that has been been passed through
    \code{data} argument.}
}
\description{
Given a data frame, a predictor (\code{IV}), an outcome
(\code{DV}), a mediator (\code{M}), and a moderator (\code{Mod}) conducts a
joint-significant test for moderated mediation (see Yzerbyt, Muller,
Batailler, & Judd, 2018).
}
\details{
This function is an helper to fit linear models needed to test a
moderated mediation model. The join-significant formal approach has been
described in Muller, Judd, & Yzerbyt (2005).

Three models will be fit:
}
\references{
Muller, D., Judd, C. M., & Yzerbyt, V. Y. (2005). When moderation
is mediated and mediation is moderated. \emph{Journal of Personality and
Social Psychology}, 89(6), 852‑863. doi: 10.1037/0022-3514.89.6.852
}
\seealso{
Other mediation models: \code{\link{mdt_simple}},
  \code{\link{mdt_within}}
}
