% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_mediation.R
\name{mdt_simple}
\alias{mdt_simple}
\title{Fit a simple mediation model}
\usage{
mdt_simple(...)
}
\arguments{
\item{data}{dataframe}

\item{IV}{IV}

\item{Mediator}{Mediator}

\item{DV}{DV}
}
\description{
Given a dataframe, an IV, a DV and a mediator, fit a mediation model
}
