% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_within.R
\name{mdt_within}
\alias{mdt_within}
\title{Fit a within-participants mediation model (long format)}
\usage{
mdt_within(data, IV, DV, M, grouping)
}
\arguments{
\item{data}{a dataframe containing the variables in the model.}

\item{IV}{an unquoted variable in the data frame which will be used
as independant variable.}

\item{DV}{an unquoted variable in the data frame which will be used
as dependant variable.}

\item{M}{an unquoted variable in the data frame which will be used
as mediator.}

\item{grouping}{an unquoted variable in the data frame which will be used
as grouping variable.}
}
\description{
Fit a within-participants mediation model (long format)
}
