% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_within2.R
\name{mdt_within2}
\alias{mdt_within2}
\title{Fit a within-participants mediation model (wide-format input)}
\usage{
mdt_within2(data, DV_A, DV_B, M_A, M_B)
}
\arguments{
\item{data}{a dataframe containing the variables in the model.}

\item{DV_A}{a}

\item{DV_B}{a}

\item{M_A}{a}

\item{M_B}{a}
}
\description{
Fit a within-participants mediation model (wide-format input)
}
