% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_figure_options.R
\name{mediaiton_figure_options}
\alias{mediaiton_figure_options}
\title{Options for mediation figure display}
\usage{
mediaiton_figure_options(
  height = 0.75,
  width = 2,
  graph_label = NA,
  node_text_size = 12,
  edge_text_size = 12,
  color = "black",
  ranksep = 0.2,
  minlen = 3
)
}
\arguments{
\item{height}{}

\item{width}{}

\item{graph_label}{}

\item{node_text_size}{}

\item{edge_text_size}{}

\item{color}{}

\item{ranksep}{}

\item{minlen}{}
}
\value{

}
\description{
Construct option list for mediation figure display. All arguments have
defaults.
}
