% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_simple_plot.R
\name{plot.simple_mediation}
\alias{plot.simple_mediation}
\title{Plot method for simple_mediation object}
\usage{
\method{plot}{simple_mediation}(
  mediation_model,
  coef_rounding = 2,
  coef_show_stars = TRUE,
  height = 0.75,
  width = 2,
  graph_label = NA,
  node_text_size = 12,
  edge_text_size = 12,
  color = "black",
  ranksep = 0.2,
  minlen = 3
)
}
\arguments{
\item{mediation_model}{}

\item{coef_rounding}{}

\item{coef_show_stars}{}

\item{height}{}

\item{width}{}

\item{graph_label}{}

\item{node_text_size}{}

\item{edge_text_size}{}

\item{color}{}

\item{ranksep}{}

\item{minlen}{}
}
\value{

}
\description{
Plot method for simple_mediation object
}
