% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indirect_effect-methods.R
\name{print.indirect_index}
\alias{print.indirect_index}
\title{Print a summary for an indirect effect index created with \code{add_index()}
method.}
\usage{
\method{print}{indirect_index}(x, digits = 3)
}
\description{
Print a summary for an indirect effect index created with \code{add_index()}
method.
}
\author{
Cédric Batailler \email{cedric.batailler@gmail.com}
}
