% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_model-methods.R
\name{print.mediation_model}
\alias{print.mediation_model}
\title{Print a summary for a mediation model represented by a \code{mediation_model} object}
\usage{
\method{print}{mediation_model}(x, ...)
}
\description{
Print a summary for a mediation model represented by a \code{mediation_model} object
}
\author{
Cédric Batailler \email{cedric.batailler@gmail.com}
}
