% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_wide_to_long.R
\name{shape_wide_to_long}
\alias{shape_wide_to_long}
\title{Shape a data frame from wide to long format.}
\usage{
shape_wide_to_long(x, DV_A, DV_B, M_A)
}
\arguments{
\item{DV_A}{An unquoted variable name in data containing DV value for A
condition of IV.}

\item{DV_B}{An unquoted variable name in data containing DV value for B
condition of IV.}

\item{M_A}{An unquoted variable name in data containing DV value for A
condition of mediator.}

\item{data}{A \code{data.frame}}

\item{M_B}{An unquoted variable name in data containing DV value for B
condition of mediator.}
}
\value{
data A \code{data.frame} in a long format.
}
\description{
An helper to reshape a wide formatted data.frame to a long one.
\code{\link{mdt_within}} requiere a \code{data.frame} with one line by IV
condition. \code{reshape_wide_to_long} helps you to go from a
\code{data.frame} where DV and M score would be stored in different columns
for the different IV condition to a \code{data.frame} that is supported by
\code{\link{mdt_within}}.
}
\examples{
\dontrun{
}
}
