% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{simple_mediation}
\alias{simple_mediation}
\title{Fit a simple mediation model}
\usage{
simple_mediation(...)
}
\arguments{
\item{data}{dataframe}

\item{IV}{IV}

\item{Mediator}{Mediator}

\item{DV}{DV}
}
\description{
Given a dataframe, an IV, a DV and a mediator, fit a mediation model
}
