% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standarize_variables.R
\name{standardize_variables}
\alias{standardize_variables}
\alias{standardise_variables}
\title{Standardize variables in a data set.}
\usage{
standardize_variables(.data, .cols = dplyr::everything())

standardise_variables(.data, .cols = dplyr::everything())
}
\arguments{
\item{.data}{A data frame containing the variables to standardize.}

\item{.cols}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Columns to standardize.}
}
\value{
A data frame with the standardized columns.
}
\description{
\code{standardize_variables()} standardizes the selected columns in a data.frame
using \code{\link[base:scale]{scale}}. This function overwrites
the column to be scaled.

\code{standardize_variables()} and \code{standardise_variables()} are synonyms.
}
\examples{
ho_et_al \%>\%
  standardize_variables(sdo)

ho_et_al \%>\%
  standardize_variables(c(sdo, linkedfate))

}
