% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_index.R
\name{add_index}
\alias{add_index}
\title{Add mediation index to a fitted mediation model}
\usage{
add_index(mediation_model, iter = 5000, alpha = 0.05, stage = NULL)
}
\arguments{
\item{mediation_model}{A mediation model fitted with \code{mediation_model}
method.}

\item{iter}{number of iteration for the Monte Carlo CI method.}

\item{alpha}{Alpha threshold to use when computing Monte Carlo CI.}

\item{stage}{Stage on which you want to compute conditional index for
moderated mediation models.}
}
\value{
An object of the same class as \code{mediation_model}, but with
confidence interval for the indirect effect estimation added for later
use.
}
\description{
Add mediation index to a fitted mediation model.
}
\examples{
\dontrun{
my_model <- add_ci(my_model)
}
}
