% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_contrast.R
\name{build_contrast}
\alias{build_contrast}
\title{Build a contrast code from character vector}
\usage{
build_contrast(vector, cond_a, cond_b)
}
\arguments{
\item{vector}{A character vector.}

\item{cond_a}{A character string to be coded -0.5.}

\item{cond_b}{A character string to be coded  0.5.}
}
\value{
A numeric vector.
}
\description{
This helper function helps you build a contrast code from a
character variable. It is useful when you need to recode a two-categories
character variable to a numeric one.
}
\details{
On \code{R}, \code{\link{lm}} method supports factor and characters
variable by dummy coding them. Dummy coding can make interpretation of
regression coefficent hard or at least harder than contrast coding.
Contrast-coded-variable coefficients interpretation is particularly useful
when conducting a joint-significance test.
}
\examples{
  data(ho_et_al)

  ho_et_al$condition_contrast <- build_contrast(ho_et_al$condition,
                                                "Low discrimination",
                                                "High discrimination")

 head(ho_et_al)

}
\seealso{
\code{\link{scale}} for centering continuous numeric variable.
}
