% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_contrast.R
\name{build_contrast}
\alias{build_contrast}
\title{build a contrast code from character vector}
\usage{
build_contrast(vector, cond_a, cond_b)
}
\arguments{
\item{vector}{a character vector.}

\item{cond_a}{a character string to be recoded -0.5}

\item{cond_b}{a character string to be recoded  0.5}
}
