% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_indirect_effect_for.R
\name{compute_indirect_effect_for}
\alias{compute_indirect_effect_for}
\title{Compute the indirect effect index for a specific value of the moderator}
\usage{
compute_indirect_effect_for()
}
\description{
#' @export
}
