% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_models.R
\name{extract_models}
\alias{extract_models}
\title{Extract models from a mediation_model object}
\usage{
extract_models(mediation_model, step = NULL)
}
\arguments{
\item{mediation_model}{An object of class \code{mediation_model}.}

\item{step}{An integer or a string corresponding to the model
to extract (optional).}
}
\value{
Either a list of \code{lm} objects or an \code{lm} object.
}
\description{
When conducting a joint significant test, different models are
fitted to the data. This function helps you access the models that have been
used in an object of class \code{mediation_model}.
}
\examples{
\dontrun{
extract_models(my_model)
}

}
\seealso{
\code{\link{extract_models}} to access to models used for
joint-significant test as a data frame.
}
