% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_models.R
\name{extract_models}
\alias{extract_models}
\title{Helper to extract models from a mediation object}
\usage{
extract_models(mediation_model)
}
\arguments{
\item{mediation_model}{A mediation model fitted with \code{mediation_model}
method.}
}
\value{
A list of \code{lm} objects.
}
\description{
With joint significant approach, different models are fitted to the data.
This function helps you access the models that have been fit when using
\code{mediation_model()}.
}
