% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_models.R
\name{extract_tidy_models}
\alias{extract_tidy_models}
\title{Helper to extract tidy models from a mediation object}
\usage{
extract_tidy_models(mediation_model)
}
\arguments{
\item{mediation_model}{A mediation model fitted with \code{mediation_model}
method.}
}
\value{
A data.frame with models information.
}
\description{
When conducting a joint significant test, different models are
fitted to the data. This function helps you access the models that have been
used in an object of class \code{mediation_model}.
}
