% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_moderated.R
\name{mdt_moderated}
\alias{mdt_moderated}
\title{Fit a moderated mediation model}
\usage{
mdt_moderated(...)
}
\arguments{
\item{data}{dataframe}

\item{IV}{IV}

\item{M}{Mediator}

\item{DV}{DV}

\item{Moderator}{Moderator}
}
\description{
Given a dataframe, an IV, a DV, a mediator and a moderator, fit a moderated
mediation model.
}
\references{
Muller, D., Judd, C. M., & Yzerbyt, V. Y. (2005). When moderation
is mediated and mediation is moderated. Journal of Personality and Social
Psychology, 89(6), 852‑863. doi: 10.1037/0022-3514.89.6.852
}
