% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_moderated.R
\name{mdt_moderated}
\alias{mdt_moderated}
\title{Fit a moderated mediation model}
\usage{
mdt_moderated(...)
}
\arguments{
\item{data}{dataframe}

\item{IV}{IV}

\item{Mediator}{Mediator}

\item{DV}{DV}
}
\description{
Given a dataframe, an IV, a DV, a mediator and a moderator, fit a moderated
mediation model.
}
