% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_simple.R
\name{mdt_simple}
\alias{mdt_simple}
\title{Simple mediation model}
\usage{
mdt_simple(...)
}
\arguments{
\item{data}{a dataframe containing the variables in the model.}

\item{IV}{an unquoted variable in the data frame which will be used
as independant variable.}

\item{M}{an unquoted variable in the data frame which will be used
as mediator.}

\item{DV}{an unquoted variable in the data frame which will be used
as dependant variable.}
}
\description{
\code{mdt_simple} fits a simple mediation model using joint significance
method (Judd, Yzerbyt, & Muller, 2014).
}
\details{
Given a dataframe, an IV, a DV and a mediator, fits a mediation model.
}
\references{
Judd, C. M., Yzerbyt, V. Y., & Muller, D. (2014). Mediation
and moderation. Handbook of research methods in
social and personality psychology, 2, 653–676.
}
