% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_simple.R
\name{mdt_simple}
\alias{mdt_simple}
\title{Fit a simple mediation model}
\usage{
mdt_simple(...)
}
\arguments{
\item{data}{a dataframe containing the variables in the model.}

\item{IV}{an unquoted variable in the data frame which will be used
as independant variable.}

\item{M}{an unquoted variable in the data frame which will be used
as mediator.}

\item{DV}{an unquoted variable in the data frame which will be used
as dependant variable.}
}
\description{
Given a dataframe, an IV, a DV and a mediator, fit a mediation model
}
