% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_within.R
\name{mdt_within}
\alias{mdt_within}
\title{Fit a within-participant mediation model (long format)}
\usage{
mdt_within(data, IV, DV, M, grouping, default_coding = TRUE)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{IV}{an unquoted variable in the data frame which will be used
as independant variable.}

\item{DV}{an unquoted variable in the data frame which will be used
as dependant variable.}

\item{M}{an unquoted variable in the data frame which will be used
as mediator.}

\item{grouping}{an unquoted variable in the data frame which will be used
as grouping variable.}

\item{default_coding}{should the variable coding be the default? Defaults to
\code{TRUE}.}
}
\description{
Given a data frame, an IV, a DV, and a mediator, fits
a within-participant mediation model.
}
\section{Coding}{

Within-participant mediation model's underlying models uses difference
scores as DV. Because the function input does not allow the user
to specify how the difference scores should be computed, \code{mdt_within}
has to figure the coding by itself.

\code{mdt_within}'s default behavior is to choose a coding where the total
effect (the effect of \eqn{X} on \eqn{Y}) will be positive and compute
other difference scores accordingly.

User can change choose to have a negative total effect by using the
\code{default_coding} argument.
}

