% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_within2.R
\name{mdt_within2}
\alias{mdt_within2}
\title{Fit a within-subject mediation model}
\usage{
mdt_within2(data, DV_A, DV_B, M_A, M_B)
}
\arguments{
\item{data}{a dataframe containing the variables in the model.}

\item{M_B}{}
}
\description{
Fit a within-subject mediation model
}
