% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdt_simple_plot.R
\name{plot.simple_mediation}
\alias{plot.simple_mediation}
\title{Plot method for simple_mediation object}
\usage{
\method{plot}{simple_mediation}(
  x,
  coef_rounding = 2,
  coef_show_stars = TRUE,
  figure_options = mediaiton_figure_options(),
  ...
)
}
\arguments{
\item{x}{A mediation model created with \code{mdt_simple}.}

\item{coef_rounding}{}

\item{coef_show_stars}{}

\item{...}{}
}
\value{

}
\description{
Plot method for simple_mediation object
}
