% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indirect_effect-methods.R
\name{print.indirect_index}
\alias{print.indirect_index}
\title{Print a summary for an indirect effect index created with \code{add_index()}
method.}
\usage{
\method{print}{indirect_index}(x, digits = 3)
}
\arguments{
\item{x}{An object of class \code{indirect_index}.}

\item{digits}{How many significant digits are to be used for numerics.}
}
\description{
Print a summary for an indirect effect index created with
\code{add_index()} method.
}
