% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_model_print.R
\name{print.mediation_model}
\alias{print.mediation_model}
\title{Print method for object of class \code{mediation_model}}
\usage{
\method{print}{mediation_model}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{mediation_model}.}

\item{digits}{How many significant digits are to be used for numerics.}

\item{...}{Further arguments.}
}
\description{
Print a summary for a mediation model represented by a
\code{mediation_model} object.
}
