% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_model-methods.R
\name{print.mediation_model}
\alias{print.mediation_model}
\title{Print a summary for a mediation model represented by a \code{mediation_model}
object.}
\usage{
\method{print}{mediation_model}(x, digits = 3)
}
\arguments{
\item{x}{An object of class \code{mediation_model}.}

\item{digits}{How many significant digits are to be used for numerics.}
}
\description{
Print a summary for a mediation model represented by a
\code{mediation_model} object.
}
